Fixing implicite double -> int conversions
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@1081 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ examples/realtime/randomplot.cpp
@@ -37,7 +37,7 @@ public:
         QwtScaleWidget *scaleWidget = plot()->axisWidget( yAxis() );
         QwtScaleDraw *sd = scaleWidget->scaleDraw();
 
-        int minExtent = 0;
+        double minExtent = 0.0;
         if ( zoomRectIndex() > 0 )
         {
             // When scrolling in vertical direction
